// arcadia.backoffice.fe/src/modules/dashboard/types.ts
export enum ActiveSessionsBreakdownColor {
  reBuy = '#737374',
  observing = '#32c5ff',
  queuing = '#e02020',
  betBehind = '#f7b500',
  inPlay = '#07af08',
}

export enum ActiveSessionsBreakdownLabel {
  observing = 'Observing',
  queuing = 'Queuing',
  betBehind = 'Bet Behind',
  inPlay = 'In Play',
  reBuy = 'ReBuy',
}

export enum ActiveAndNewPlayersLabel {
  active = 'Active',
  new = 'New',
}

export enum ActiveAndNewPlayersColor {
  active = '#32c5ff',
  new = '#e02020',
}

export enum BettingActivityLabel {
  rounds = 'Rounds',
  bets = 'Bets',
  wins = 'Wins',
}

export enum BettingActivityColor {
  rounds = '#07af08',
  bets = '#f7b500',
  wins = '#32c5ff',
}

export interface IActiveSessionsBreakdownChartItem {
  name: ActiveSessionsBreakdownLabel
  color: ActiveSessionsBreakdownColor
  value: number
}

export type IActiveSessionsBreakdownChartData = IActiveSessionsBreakdownChartItem[];